+++ lib/html.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library web_socket_channel.html;
-
 import 'dart:async';
 import 'dart:html';
 import 'dart:typed_data';
@@ -20,11 +18,14 @@ class HtmlWebSocketChannel extends StreamChannelMixin
   /// The underlying `dart:html` [WebSocket].
   final WebSocket _webSocket;
 
+  @override
   String get protocol => _webSocket.protocol;
 
+  @override
   int get closeCode => _closeCode;
   int _closeCode;
 
+  @override
   String get closeReason => _closeReason;
   String _closeReason;
 
@@ -44,10 +45,12 @@ class HtmlWebSocketChannel extends StreamChannelMixin
   /// [_controller.local.stream].
   String _localCloseReason;
 
+  @override
   Stream get stream => _controller.foreign.stream;
   final _controller =
       StreamChannelController(sync: true, allowForeignErrors: false);
 
+  @override
   WebSocketSink get sink => _sink;
   WebSocketSink _sink;
 
@@ -85,7 +88,7 @@ class HtmlWebSocketChannel extends StreamChannelMixin
     // and that once it is no open or message events will be emitted.
     _webSocket.onError.first.then((_) {
       _controller.local.sink
-          .addError(WebSocketChannelException("WebSocket connection failed."));
+          .addError(WebSocketChannelException('WebSocket connection failed.'));
       _controller.local.sink.close();
     });
 
@@ -131,6 +134,7 @@ class _HtmlWebSocketSink extends DelegatingStreamSink implements WebSocketSink {
       : _channel = channel,
         super(channel._controller.foreign.sink);
 
+  @override
   Future close([int closeCode, String closeReason]) {
     _channel._localCloseCode = closeCode;
     _channel._localCloseReason = closeReason;
@@ -142,10 +146,10 @@ class _HtmlWebSocketSink extends DelegatingStreamSink implements WebSocketSink {
 /// messages.
 class BinaryType {
   /// Tells the channel to emit binary messages as [Blob]s.
-  static const blob = BinaryType._("blob", "blob");
+  static const blob = BinaryType._('blob', 'blob');
 
   /// Tells the channel to emit binary messages as [Uint8List]s.
-  static const list = BinaryType._("list", "arraybuffer");
+  static const list = BinaryType._('list', 'arraybuffer');
 
   /// The name of the binary type, which matches its variable name.
   final String name;
@@ -155,5 +159,6 @@ class BinaryType {
 
   const BinaryType._(this.name, this.value);
 
+  @override
   String toString() => name;
 }