+++ lib/io.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library web_socket_channel.io;
-
 import 'dart:async';
 import 'dart:io';
 
@@ -23,10 +21,14 @@ class IOWebSocketChannel extends StreamChannelMixin
   /// `null` until the [WebSocket.connect] future completes.
   WebSocket _webSocket;
 
+  @override
   String get protocol => _webSocket?.protocol;
+  @override
   int get closeCode => _webSocket?.closeCode;
+  @override
   String get closeReason => _webSocket?.closeReason;
 
+  @override
   /// Future indicating if the connection has been established.
   /// It completes on successful connection to the websocket.
   Future<void> get ready => _readyCompleter?.future;
@@ -34,7 +36,9 @@ class IOWebSocketChannel extends StreamChannelMixin
   /// Completer for [ready].
   final Completer _readyCompleter;
 
+  @override
   final Stream stream;
+  @override
   final WebSocketSink sink;
 
   // TODO(nweiz): Add a compression parameter after the initial release.
@@ -111,6 +115,7 @@ class _IOWebSocketSink extends DelegatingStreamSink implements WebSocketSink {
       : _webSocket = webSocket,
         super(webSocket);
 
+  @override
   Future close([int closeCode, String closeReason]) =>
       _webSocket.close(closeCode, closeReason);
 }