+++ test/io_test.dart
@@ -17,19 +17,19 @@ void main() {
     if (server != null) await server.close();
   });
 
-  test("communicates using existing WebSockets", () async {
-    server = await HttpServer.bind("localhost", 0);
+  test('communicates using existing WebSockets', () async {
+    server = await HttpServer.bind('localhost', 0);
     server.transform(WebSocketTransformer()).listen((webSocket) {
       var channel = IOWebSocketChannel(webSocket);
-      channel.sink.add("hello!");
+      channel.sink.add('hello!');
       channel.stream.listen((request) {
-        expect(request, equals("ping"));
-        channel.sink.add("pong");
-        channel.sink.close(5678, "raisin");
+        expect(request, equals('ping'));
+        channel.sink.add('pong');
+        channel.sink.close(5678, 'raisin');
       });
     });
 
-    var webSocket = await WebSocket.connect("ws://localhost:${server.port}");
+    var webSocket = await WebSocket.connect('ws://localhost:${server.port}');
     var channel = IOWebSocketChannel(webSocket);
 
     expect(channel.ready, completes);
@@ -37,31 +37,31 @@ void main() {
     var n = 0;
     channel.stream.listen((message) {
       if (n == 0) {
-        expect(message, equals("hello!"));
-        channel.sink.add("ping");
+        expect(message, equals('hello!'));
+        channel.sink.add('ping');
       } else if (n == 1) {
-        expect(message, equals("pong"));
+        expect(message, equals('pong'));
       } else {
-        fail("Only expected two messages.");
+        fail('Only expected two messages.');
       }
       n++;
     }, onDone: expectAsync0(() {
       expect(channel.closeCode, equals(5678));
-      expect(channel.closeReason, equals("raisin"));
+      expect(channel.closeReason, equals('raisin'));
     }));
   });
 
-  test(".connect communicates immediately", () async {
-    server = await HttpServer.bind("localhost", 0);
+  test('.connect communicates immediately', () async {
+    server = await HttpServer.bind('localhost', 0);
     server.transform(WebSocketTransformer()).listen((webSocket) {
       var channel = IOWebSocketChannel(webSocket);
       channel.stream.listen((request) {
-        expect(request, equals("ping"));
-        channel.sink.add("pong");
+        expect(request, equals('ping'));
+        channel.sink.add('pong');
       });
     });
 
-    var channel = IOWebSocketChannel.connect("ws://localhost:${server.port}");
+    var channel = IOWebSocketChannel.connect('ws://localhost:${server.port}');
 
     expect(channel.ready, completes);
 
@@ -69,66 +69,65 @@ void main() {
 
     channel.stream.listen(
         expectAsync1((message) {
-          expect(message, equals("pong"));
-          channel.sink.close(5678, "raisin");
+          expect(message, equals('pong'));
+          channel.sink.close(5678, 'raisin');
         }, count: 1),
         onDone: expectAsync0(() {}));
   });
 
-  test(".connect communicates immediately using platform independent api",
+  test('.connect communicates immediately using platform independent api',
       () async {
-    server = await HttpServer.bind("localhost", 0);
+    server = await HttpServer.bind('localhost', 0);
     server.transform(WebSocketTransformer()).listen((webSocket) {
       var channel = IOWebSocketChannel(webSocket);
       channel.stream.listen((request) {
-        expect(request, equals("ping"));
-        channel.sink.add("pong");
+        expect(request, equals('ping'));
+        channel.sink.add('pong');
       });
     });
 
     var channel =
-        WebSocketChannel.connect(Uri.parse("ws://localhost:${server.port}"));
-
-    channel.sink.add("ping");
+        WebSocketChannel.connect(Uri.parse('ws://localhost:${server.port}'));
+    channel.sink.add('ping');
 
     channel.stream.listen(
         expectAsync1((message) {
-          expect(message, equals("pong"));
-          channel.sink.close(5678, "raisin");
+          expect(message, equals('pong'));
+          channel.sink.close(5678, 'raisin');
         }, count: 1),
         onDone: expectAsync0(() {}));
   });
 
-  test(".connect with an immediate call to close", () async {
-    server = await HttpServer.bind("localhost", 0);
+  test('.connect with an immediate call to close', () async {
+    server = await HttpServer.bind('localhost', 0);
     server.transform(WebSocketTransformer()).listen((webSocket) {
       expect(() async {
         var channel = IOWebSocketChannel(webSocket);
         await channel.stream.drain();
         expect(channel.closeCode, equals(5678));
-        expect(channel.closeReason, equals("raisin"));
+        expect(channel.closeReason, equals('raisin'));
       }(), completes);
     });
 
-    var channel = IOWebSocketChannel.connect("ws://localhost:${server.port}");
+    var channel = IOWebSocketChannel.connect('ws://localhost:${server.port}');
     expect(channel.ready, completes);
-    await channel.sink.close(5678, "raisin");
+    await channel.sink.close(5678, 'raisin');
   });
 
-  test(".connect wraps a connection error in WebSocketChannelException",
+  test('.connect wraps a connection error in WebSocketChannelException',
       () async {
-    server = await HttpServer.bind("localhost", 0);
+    server = await HttpServer.bind('localhost', 0);
     server.listen((request) {
       request.response.statusCode = 404;
       request.response.close();
     });
 
-    var channel = IOWebSocketChannel.connect("ws://localhost:${server.port}");
+    var channel = IOWebSocketChannel.connect('ws://localhost:${server.port}');
     expect(channel.stream.drain(),
         throwsA(TypeMatcher<WebSocketChannelException>()));
   });
 
-  test(".protocols fail", () async {
+  test('.protocols fail', () async {
     var passedProtocol = 'passed-protocol';
     var failedProtocol = 'failed-protocol';
     var selector = (List<String> receivedProtocols) => passedProtocol;
@@ -139,13 +138,13 @@ void main() {
           throwsException);
     });
 
-    var channel = IOWebSocketChannel.connect("ws://localhost:${server.port}",
+    var channel = IOWebSocketChannel.connect('ws://localhost:${server.port}',
         protocols: [failedProtocol]);
     expect(channel.stream.drain(),
         throwsA(TypeMatcher<WebSocketChannelException>()));
   });
 
-  test(".protocols pass", () async {
+  test('.protocols pass', () async {
     var passedProtocol = 'passed-protocol';
     var selector = (List<String> receivedProtocols) => passedProtocol;
 
@@ -157,7 +156,7 @@ void main() {
       await webSocket.close();
     });
 
-    var channel = IOWebSocketChannel.connect("ws://localhost:${server.port}",
+    var channel = IOWebSocketChannel.connect('ws://localhost:${server.port}',
         protocols: [passedProtocol]);
     await channel.stream.drain();
     expect(channel.protocol, passedProtocol);